Added new preferences page
